+++ example/lib/shared/decoration/chest.dart
@@ -110,10 +110,4 @@ class Chest extends GameDecoration with TapGesture {
       ),
     );
   }
-
-  @override
-  void onTapDown(int pointer, Vector2 position) {}
-
-  @override
-  void onTapUp(int pointer, Vector2 position) {}
 }